took out collapsed hamburger option on mobile
This feature was getting in the way of the view of the list of pokemon. When user searches for pokemon, search bar does not hide away after hitting enter. I opted to take this out vs. collapsing hamburger menu with nested search bar because I wanted mobile version to be simple and straight forward. No need for extra links like "Home", etc.

Also, d-none to GitHub link on mobile